refactor function and tests

+++ src/controllers/lesson.controller.js
@@ -1,9 +1,5 @@
 const Lessons = require("../models/lesson.model");
 
-const helloLesson = (req, res) => {
-  res.status(200).json({ message: "Hello from the lesson controller!" });
-};
-
 const getLessons = async (req, res) => {
   try {
     const lessons = await Lessons.find();
@@ -25,32 +21,20 @@ const getSingleLesson = async (req, res) => {
 const postLesson = async (req, res) => {
   try {
     const body = req.body;
-
-    if (!body.title || body.title === "") {
+    const error = checkLesson(body);
+    if (error !== null) {
       return res.status(400).json({
-        message: "Title is required",
+        message: error,
       });
-    }
-
-    if (!body.content || body.content === "") {
-      return res.status(400).json({
-        message: "Content is required",
+    } else {
+      const newLesson = new Lessons({
+        title: req.body.title,
+        content: req.body.content,
+        isPublished: req.body.isPublished,
       });
+      await newLesson.save();
+      res.status(201).json(newLesson);
     }
-
-    if (body.isPublished === undefined) {
-      return res.status(400).json({
-        message: "Lesson must be published or not",
-      });
-    }
-
-    const newLesson = new Lessons({
-      title: req.body.title,
-      content: req.body.content,
-      isPublished: req.body.isPublished,
-    });
-    await newLesson.save();
-    res.status(201).json(newLesson);
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -61,40 +45,35 @@ const postLesson = async (req, res) => {
 const updateLesson = async (req, res) => {
   try {
     const body = req.body;
-
-    if (!body.title || body.title === "") {
+    const error = checkLesson(body);
+    if (error !== null) {
       return res.status(400).json({
-        message: "Title is required",
+        message: error,
       });
+    } else {
+      const lesson = await Lessons.findById(req.params.id);
+      lesson.title = req.body.title;
+      lesson.content = req.body.content;
+      lesson.isPublished = req.body.isPublished;
+      await lesson.save();
+      res.status(200).json({ message: "Lesson updated" });
     }
-
-    if (!body.content || body.content === "") {
-      return res.status(400).json({
-        message: "Content is required",
-      });
-    }
-
-    if (body.isPublished === undefined) {
-      return res.status(400).json({
-        message: "Lesson must be published or not",
-      });
-    }
-
-    const lesson = await Lessons.findById(req.params.id);
-    lesson.title = req.body.title;
-    lesson.content = req.body.content;
-    lesson.isPublished = req.body.isPublished;
-    await lesson.save();
-    res.status(200).json({ message: "Lesson updated" });
   } catch (error) {
     res.status(500).json({
       message: error.message,
     });
   }
-}
+};
 
 const deleteLesson = async (req, res) => {
   try {
+    const lesson = await Lessons.findById(req.params.id);
+    if (!lesson) {
+      return res.status(404).json({
+        message: "Lesson not found",
+      });
+    }
+
     await Lessons.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: "Lesson deleted" });
   } catch (error) {
@@ -102,10 +81,25 @@ const deleteLesson = async (req, res) => {
       message: error.message,
     });
   }
-}
+};
+
+const checkLesson = (lesson) => {
+  if (!lesson.title || lesson.title === "") {
+    return "Title is required";
+  }
+
+  if (!lesson.content || lesson.content === "") {
+    return "Content is required";
+  }
+
+  if (lesson.isPublished === undefined) {
+    return "Lesson must be published or not";
+  }
+
+  return null;
+};
 
 module.exports = {
-  helloLesson,
   getLessons,
   getSingleLesson,
   postLesson,
